Moved tests around and added error message
