+++ src/IntegrationTests/java/uk/gov/hmcts/sscs/postcode/PostCodeEndpointsIt.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sscs;
+package uk.gov.hmcts.sscs.postcode;
 
 import static org.junit.Assert.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
