+++ src/test/java/uk/gov/hmcts/sscs/service/AirLookupServiceTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Before;
 import org.junit.Test;
 
-public class AirGroupLookupServiceTest {
+public class AirLookupServiceTest {
     AirLookupService airLookupService;
 
     @Before
